backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic[email]==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.9
